requirements.txt
ccxt
pandas
numpy
TA-Lib
APScheduler
aiohttp
openpyxl
SQLAlchemy
python-dotenv
PyYAML
loguru
python-telegram-bot
discord.py
fastapi
uvicorn[standard]
pydantic
requests
httpx 